+++ scripts/lib/utilities/macroUtils.js
@@ -1,4 +1,5 @@
 import {custom} from '../../events/custom.js';
 export let macroUtils = {
-    registerMacros: custom.registerMacros
+    registerMacros: custom.registerMacros,
+    getMacro: custom.getMacro
 };